refactor(segments): Move segment speed utilities into their own utils file since they are now used i
