+++ lib/components/modification/segment-speeds.js
@@ -6,6 +6,7 @@ import Collapsible from '../collapsible'
 import DeepEqual from '../deep-equal'
 import {Group, Input, Number as InputNumber} from '../input'
 import {toPrecision} from '../../utils/math'
+import {getAverageSpeedOfSegments, getTravelTimeMinutes} from '../../utils/segments'
 
 export default class SegmentSpeeds extends DeepEqual {
   static propTypes = {
@@ -182,31 +183,5 @@ export default class SegmentSpeeds extends DeepEqual {
         </Collapsible>
       </div>
     )
-  }
-}
-
-export function getAverageSpeedOfSegments ({segmentDistances, segmentSpeeds}) {
-  const totalDistance = sum(segmentDistances)
-  const weightedSpeeds = segmentSpeeds.map((s, i) => (s * segmentDistances[i]))
-  return weightedSpeeds.reduce((total, speed) => (total + speed), 0) / totalDistance
-}
-
-function getTravelTimeMinutes ({
-  dwellTime, // seconds
-  numberOfStops,
-  segmentDistances, // km
-  segmentSpeeds // kph
-}) {
-  if (numberOfStops > 0) {
-    const segmentTimes = segmentDistances.map(
-      (segmentDistance, i) => (segmentDistance / (segmentSpeeds[i] || DEFAULT_SEGMENT_SPEED)))
-    const totalMovingTime = sum(segmentTimes) * 60
-
-    // add the dwell times
-    const dwellMinutes = dwellTime / 60
-    const dwellingTime = numberOfStops * dwellMinutes
-    return totalMovingTime + dwellingTime
-  } else {
-    return 0
   }
 }
