+++ lib/components/report/reroute.js
@@ -10,9 +10,9 @@ import Distance from './distance'
 import MiniMap from './mini-map'
 import RerouteLayer from '../scenario-map/reroute-layer'
 import Speed from './speed'
-import getStops from '../../utils/get-stops'
 import messages from '../../utils/messages'
 import {getPatternsForModification} from '../../utils/patterns'
+import getStops from '../../utils/get-stops'
 import {getAverageSpeedOfSegments} from '../../utils/segments'
 
 export default class Reroute extends Component {
